Added commands for music
Added queue, pause, resume, skip, clear, and playlist commands

+++ bot.py
@@ -29,6 +29,7 @@ client = discord.Client(intents=discord.Intents.all())
 # Keep track of the voice client
 voice_client = None
 filename = None
+playlist = []
 
 #endregion
 
@@ -36,7 +37,7 @@ filename = None
 
 # To be able to delete the music file when the bot is closed
 def delete_file_on_exit():
-    if os.path.exists(filename):
+    if filename and os.path.exists(filename):
         time.sleep(1)
         os.remove(filename)
 
@@ -77,7 +78,7 @@ def generate_image(message):
     return response.data[0]['url']
 
 # Play a video on YouTube
-def get_youtube_song(message, query):
+def get_youtube_song(query):
     # Use the YouTube Data API to search for videos that match the query
         youtube = googleapiclient.discovery.build("youtube", "v3", developerKey=YOUTUBE_API_KEY)
         request = youtube.search().list(
@@ -90,6 +91,30 @@ def get_youtube_song(message, query):
         )
         response = request.execute()
         return response
+    
+# Queue a song to be played
+async def queue_song(message, query):
+    global playlist
+    try:
+        if message.author.voice is None:
+            await message.channel.send('You are not in a voice channel. Please join a voice channel and try again.')
+            return
+        # Get the search query from the message content
+        response = get_youtube_song(query)
+        # Get the first video from the search results
+        if response["items"]:
+            video_id = response["items"][0]["id"]["videoId"]
+        else:
+            print("No results found")
+            await message.channel.send("Could not find a video with that name. Please try again.")
+            return
+        # Append the video ID to the playlist
+        playlist.append(video_id)
+        await message.channel.send("Song added to the playlist.")
+    except Exception as e:
+        print(e)
+        await message.channel.send("Could not add the song to the playlist. Please try again.")
+        return
 
 # Return all the bot's commands
 def get_commands():
@@ -101,7 +126,12 @@ def get_commands():
     `!stock-ticker` - I will return the ticker symbol of the stock you request\n\
     `!stock-info` - I will return the stock information of the stock you request\n\
     `!play` - I will play the song you request (It may help to put the title inside "" quotations)\n\
-    `!stop` - I will stop playing the song\n\
+    `!stop` - I will stop playing the song and clear the music queue\n\
+    `!queue` - I will queue the song and it will be played after the one playing is done\n\
+    `!clear` - I will clear the music queue\n\
+    `!playlist` - I will tell you how many songs are in queue\n\
+    `!pause` - I will pause the song\n\
+    `!resume` - I will resume the song\n\
     I will also make sure to greet you when you join the server!'
     
 #endregion
@@ -116,6 +146,7 @@ async def on_message(message):
     
     global voice_client
     global filename
+    global playlist
     
     if '!help' in content:
         await message.channel.send(get_commands())
@@ -156,45 +187,117 @@ async def on_message(message):
                 await message.channel.send(file=file)
         if(os.path.exists('stocks.csv')):
             os.remove('stocks.csv')
-    if message.content.startswith("!play"):
-        if voice_client is not None and voice_client.is_connected():
-            await message.channel.send(f"I am already playing a song in the voice channel \"{voice_client.channel}\". Please use the `!stop` command to stop the current song.")
-            return
-        if message.author.voice is None:
-            await message.channel.send('You are not in a voice channel. Please join a voice channel and try again.')
-            return
-        # Get the search query from the message content
-        response = get_youtube_song(message, content[6:])
-        # Get the first video from the search results
-        if response["items"]:
-            video_id = response["items"][0]["id"]["videoId"]
-        else:
-            print("No results found")
-            await message.channel.send("Could not find a video with that name. Please try again.")
-            return
-        # Get the voice channel the user is in
-        voice_channel = message.author.voice.channel
-        # Connect to the voice channel\
-        if voice_client is None or voice_client.is_connected() == False:
-            voice_client = await voice_channel.connect()
-        # Use pytube to download the audio from the YouTube video
-        video = pytube.YouTube(f"https://www.youtube.com/watch?v={video_id}", use_oauth=True, allow_oauth_cache=True).streams.filter(only_audio=True).first()
-        video.download(".")
-        filename = video.default_filename
-        await message.channel.send(f"Playing `{filename[0:-4]}` in voice channel \"{voice_client.channel}\"")
-        
-        # Create a discord.FFmpegPCMAudio object to play the audio
-        audio = discord.FFmpegPCMAudio(f"{filename}")
+    if message.content.startswith("!play") and not message.content.startswith("!playlist"):
+        try:
+            if voice_client is not None and voice_client.is_connected() and voice_client.is_playing():
+                if voice_client.is_paused():
+                    await message.channel.send(f"There is already a song that is paused in the voice channel \"{voice_client.channel}\". Please use the `!resume` command to resume the song or the `!queue` command to add it to the playlist. If you wish to stop the music and clear the playlist, use the `!stop` command.")
+                else:
+                    await message.channel.send(f"I am already playing a song in the voice channel \"{voice_client.channel}\". Please use the `!stop` command to stop the current song or use the `!queue` command to add it to the playlist.")
+                return
+            if message.author.voice is None:
+                await message.channel.send('You are not in a voice channel. Please join a voice channel and try again.')
+                return
+            if voice_client and voice_client.is_paused():
+                await message.channel.send('I am currently paused. Please use the `!resume` command to resume the song.')
+                return
+            # Check if the playlist is empty
+            if not playlist and content[6:] == "":
+                await message.channel.send("Please enter a song name.")
+                return
+            if not playlist:
+                await queue_song(message, content[6:])
+            while playlist:
+                # Get the voice channel the user is in
+                voice_channel = message.author.voice.channel
+                # Connect to the voice channel\
+                if voice_client is None or voice_client.is_connected() == False:
+                    voice_client = await voice_channel.connect()
+                # If the player is paused using the command !pause, I want this to wait until the !resume command is used
+                while voice_client.is_paused() and playlist:
+                    await asyncio.sleep(1)
+                time.sleep(2)
+                video_id = playlist.pop(0)
+                # Use pytube to download the audio from the YouTube video
+                video = pytube.YouTube(f"https://www.youtube.com/watch?v={video_id}", use_oauth=True, allow_oauth_cache=True).streams.filter(only_audio=True).first()
+                video.download(".")
+                filename = video.default_filename
+                await message.channel.send(f"Playing `{filename[0:-4]}` in voice channel \"{voice_client.channel}\"")
+                
+                # Create a discord.FFmpegPCMAudio object to play the audio
+                audio = discord.FFmpegPCMAudio(f"{filename}")
 
-        # Play the audio
-        voice_client.play(audio)
-        # Wait for the audio to finish playing
-        while (voice_client is not None) and voice_client.is_playing():
-            await asyncio.sleep(1)
-        # Disconnect from the voice channel
-        if (voice_client is not None) and (voice_client.is_playing() == False or voice_client.is_connected()):
-            await voice_client.disconnect()
-            voice_client = None
+                # Play the audio
+                voice_client.play(audio)
+                # Wait for the audio to finish playing
+                while (voice_client is not None) and voice_client.is_playing():
+                    await asyncio.sleep(1)
+                if os.path.exists(filename) and not voice_client.is_paused():
+                    os.remove(filename)
+            # Disconnect from the voice channel
+            if (voice_client is not None) and not voice_client.is_paused() and (voice_client.is_playing() == False or voice_client.is_connected()):
+                await voice_client.disconnect()
+                voice_client = None
+        except Exception as e:
+            print(e)
+            if voice_client is not None and (voice_client.is_playing() == False or voice_client.is_connected()):
+                await voice_client.disconnect()
+                voice_client = None
+            if os.path.exists(filename):
+                os.remove(filename)
+            await message.channel.send("Could not play the song. Please try again.")
+            return
+            
+    if message.content.startswith("!queue"):
+        await queue_song(message, content[7:])
+        
+    if message.content.startswith("!clear"):
+        playlist.clear()
+        await message.channel.send("The playlist has been cleared.")
+    
+    if message.content.startswith("!playlist"):
+        if not playlist:
+            await message.channel.send("The playlist is empty.")
+        else:
+            await message.channel.send(f"The playlist contains {len(playlist)} song(s).")
+    
+    if message.content.startswith("!pause"):
+        try:
+            if voice_client is not None and voice_client.is_playing():
+                voice_client.pause()
+                await message.channel.send("Music playback paused.")
+            else:
+                await message.channel.send("No song is currently playing.")
+        except Exception as e:
+            print(e)
+            await message.channel.send("Could not pause the song. Please try again.")
+            return
+        
+    if message.content.startswith("!resume"):
+        try:
+            if voice_client is not None and voice_client.is_paused():
+                voice_client.resume()
+                await message.channel.send("Music playback resumed.")
+            elif voice_client is not None and voice_client.is_playing():
+                await message.channel.send("Music playback is already resumed.")
+            else:
+                await message.channel.send("No song is currently paused.")
+        except Exception as e:
+            print(e)
+            await message.channel.send("Could not resume the song. Please try again.")
+            return
+    
+    if message.content.startswith("!skip"):
+        try:
+            if voice_client is not None and voice_client.is_playing():
+                voice_client.stop()
+                await message.channel.send("Song skipped.")
+            else:
+                await message.channel.send("No song is currently playing.")
+        except Exception as e:
+            print(e)
+            await message.channel.send("Could not skip the song. Please try again.")
+            return
         
     elif message.content == "!stop":
         if voice_client is None:
@@ -206,10 +309,17 @@ async def on_message(message):
             await voice_client.disconnect()
             voice_client = None
             
-    if (voice_client is None) and (filename is not None) and os.path.exists(filename):
+        playlist.clear()
+        await message.channel.send("Music stopped. The playlist has been cleared.")
+            
+    if (voice_client is None) and (filename is not None):
         try:
             time.sleep(1)
-            os.remove(filename)
+            current_directory = os.getcwd()
+            files = os.listdir(current_directory)
+            for file in files:
+                if file.endswith(".mp4"):
+                    os.remove(file)
         except(PermissionError):
             return
         
@@ -231,29 +341,29 @@ async def on_disconnect():
 async def on_voice_state_update(member, before, after):
     global voice_client
     if before.channel is not None and (before.channel.members == [] or before.channel.members == [client.user]):
-        if voice_client:
+        if voice_client and not (voice_client.is_paused()):
             voice_client.stop()
             # Disconnect from the voice channel
-            await before.channel.voice_client.disconnect()
-        if filename is not None and os.path.exists(filename):
+            await voice_client.disconnect()
+        if filename is not None and os.path.exists(filename) and not (voice_client.is_paused()):
             time.sleep(1)
             os.remove(filename)
     # If the member stopped playing the audio or left the voice channel
     if (before.channel is not None) and (after.channel is not None):
         if (member == client.user and (not after.channel)) or (before.channel and before.channel.id == after.channel.id):
-            if voice_client:
+            if voice_client and not (voice_client.is_paused()):
                 # Stop the audio
                 voice_client.stop()
                 # Disconnect from the voice channel
                 await voice_client.disconnect()
-            if filename is not None and os.path.exists(filename):
+            if filename is not None and os.path.exists(filename) and not (voice_client.is_paused()):
                 time.sleep(1)
                 os.remove(filename)
     if member == client.user:
         if before.channel is not None and after.channel is None:
             voice_client = None
             # The bot has left the voice channel, delete the file
-            if filename is not None and os.path.exists(filename):
+            if filename is not None and os.path.exists(filename) and voice_client and not (voice_client.is_paused()):
                 time.sleep(1)
                 os.remove(filename)
